+++ android/src/main/java/ee/forgr/capacitor_updater/CapacitorUpdater.java
@@ -748,6 +748,17 @@ public class CapacitorUpdater {
     return json;
   }
 
+  private JsonObjectRequest setRetryPolicy(JsonObjectRequest request) {
+    request.setRetryPolicy(
+      new DefaultRetryPolicy(
+        this.timeout,
+        DefaultRetryPolicy.DEFAULT_MAX_RETRIES,
+        DefaultRetryPolicy.DEFAULT_BACKOFF_MULT
+      )
+    );
+    return request;
+  }
+
   private JSObject createError(String message, VolleyError error) {
     NetworkResponse response = error.networkResponse;
     final JSObject retError = new JSObject();
@@ -824,14 +835,74 @@ public class CapacitorUpdater {
         }
       }
     );
-    request.setRetryPolicy(
-      new DefaultRetryPolicy(
-        this.timeout,
-        DefaultRetryPolicy.DEFAULT_MAX_RETRIES,
-        DefaultRetryPolicy.DEFAULT_BACKOFF_MULT
-      )
+    this.requestQueue.add(setRetryPolicy(request));
+  }
+
+  public void unsetChannel(final Callback callback) {
+    String channelUrl = this.channelUrl;
+    if (
+      channelUrl == null || "".equals(channelUrl) || channelUrl.length() == 0
+    ) {
+      Log.e(TAG, "Channel URL is not set");
+      final JSObject retError = new JSObject();
+      retError.put("message", "channelUrl missing");
+      retError.put("error", "missing_config");
+      callback.callback(retError);
+      return;
+    }
+    JSONObject json = null;
+    try {
+      json = this.createInfoObject();
+    } catch (JSONException e) {
+      Log.e(TAG, "Error unsetChannel JSONException", e);
+      e.printStackTrace();
+      final JSObject retError = new JSObject();
+      retError.put("message", "Cannot get info: " + e.toString());
+      retError.put("error", "json_error");
+      callback.callback(retError);
+      return;
+    }
+    // Building a request
+    JsonObjectRequest request = new JsonObjectRequest(
+      Request.Method.DELETE,
+      channelUrl,
+      json,
+      new Response.Listener<JSONObject>() {
+        @Override
+        public void onResponse(JSONObject res) {
+          final JSObject ret = new JSObject();
+          Iterator<String> keys = res.keys();
+          while (keys.hasNext()) {
+            String key = keys.next();
+            if (res.has(key)) {
+              try {
+                ret.put(key, res.get(key));
+              } catch (JSONException e) {
+                e.printStackTrace();
+                final JSObject retError = new JSObject();
+                retError.put(
+                  "message",
+                  "Cannot unset channel: " + e.toString()
+                );
+                retError.put("error", "response_error");
+                callback.callback(ret);
+              }
+            }
+          }
+          Log.i(TAG, "Channel unset");
+          callback.callback(ret);
+        }
+      },
+      new Response.ErrorListener() {
+        @Override
+        public void onErrorResponse(VolleyError error) {
+          callback.callback(
+            CapacitorUpdater.this.createError("Error unset channel", error)
+          );
+        }
+      }
     );
-    this.requestQueue.add(request);
+    this.requestQueue.add(setRetryPolicy(request));
   }
 
   public void setChannel(final String channel, final Callback callback) {
@@ -896,14 +967,7 @@ public class CapacitorUpdater {
         }
       }
     );
-    request.setRetryPolicy(
-      new DefaultRetryPolicy(
-        this.timeout,
-        DefaultRetryPolicy.DEFAULT_MAX_RETRIES,
-        DefaultRetryPolicy.DEFAULT_BACKOFF_MULT
-      )
-    );
-    this.requestQueue.add(request);
+    this.requestQueue.add(setRetryPolicy(request));
   }
 
   public void getChannel(final Callback callback) {
@@ -963,14 +1027,7 @@ public class CapacitorUpdater {
         }
       }
     );
-    request.setRetryPolicy(
-      new DefaultRetryPolicy(
-        this.timeout,
-        DefaultRetryPolicy.DEFAULT_MAX_RETRIES,
-        DefaultRetryPolicy.DEFAULT_BACKOFF_MULT
-      )
-    );
-    this.requestQueue.add(request);
+    this.requestQueue.add(setRetryPolicy(request));
   }
 
   public void sendStats(final String action, final String versionName) {
@@ -1008,14 +1065,7 @@ public class CapacitorUpdater {
         }
       }
     );
-    request.setRetryPolicy(
-      new DefaultRetryPolicy(
-        this.timeout,
-        DefaultRetryPolicy.DEFAULT_MAX_RETRIES,
-        DefaultRetryPolicy.DEFAULT_BACKOFF_MULT
-      )
-    );
-    this.requestQueue.add(request);
+    this.requestQueue.add(setRetryPolicy(request));
   }
 
   public BundleInfo getBundleInfo(final String id) {